--- facewarp/dingwarp/.gitignore
@@ -1,5 +1,0 @@
-.DS_Store
-.vscode
-bin/*
-build
-!*/.gitkeep
