--- facewarp/dingwarp/test/reference_points.txt
@@ -1,64 +1,0 @@
-323.81 189.64
-306.32 220.75
-298.54 254.76
-298.54 293.64
-311.18 326.68
-340.34 367.50
-375.32 392.77
-419.06 409.30
-461.82 418.04
-522.08 411.24
-572.62 394.72
-611.50 361.67
-633.85 322.80
-644.54 293.64
-646.49 258.65
-639.68 226.58
-627.05 197.42
-369.49 187.70
-383.10 187.70
-397.68 187.70
-409.34 187.70
-424.89 188.67
-506.53 188.67
-523.06 187.70
-534.72 186.73
-549.30 185.76
-562.90 187.70
-457.94 238.24
-445.30 256.71
-436.56 275.17
-429.75 289.75
-420.03 304.33
-436.56 311.13
-455.99 315.02
-477.38 310.16
-495.84 304.33
-368.52 212.00
-420.03 209.08
-431.70 238.24
-419.06 265.45
-364.63 263.51
-356.86 237.27
-497.79 239.21
-511.39 207.14
-559.02 207.14
-571.65 239.21
-559.99 265.45
-507.50 266.43
-473.49 362.64
-480.29 350.01
-500.70 349.04
-520.14 349.04
-526.94 358.76
-520.14 369.45
-500.70 371.39
-478.35 372.36
-0 0
-0 459
-0 918
-419 918
-838 918
-838 459
-838 0
-419 0
