+++ facewarp/facewarp/test/triangulation.txt
@@ -1,0 +1,118 @@
+37 21 27
+21 37 20
+36 0 17
+0 36 41
+0 63 17
+63 0 56
+0 1 56
+1 0 41
+19 36 18
+36 19 37
+1 2 56
+2 1 41
+56 2 57
+2 3 57
+3 2 40
+57 3 4
+4 3 40
+36 38 39
+38 36 37
+57 4 5
+5 4 40
+18 36 17
+57 6 59
+6 57 5
+6 5 31
+37 19 20
+57 59 58
+6 7 59
+7 6 32
+29 39 28
+39 29 30
+7 8 59
+8 7 55
+37 27 38
+59 8 9
+9 8 55
+61 9 10
+9 61 60
+59 9 60
+10 9 53
+29 34 33
+34 29 28
+61 10 11
+11 10 52
+48 33 49
+33 48 32
+61 11 12
+12 11 46
+47 27 42
+27 47 28
+61 12 13
+13 12 46
+44 24 25
+24 44 43
+61 13 14
+14 13 46
+24 43 23
+14 62 61
+62 14 15
+15 14 45
+33 30 29
+30 33 32
+62 15 16
+16 15 45
+31 40 39
+40 31 5
+17 63 18
+49 34 50
+34 49 33
+18 63 19
+11 52 51
+19 63 20
+21 63 22
+63 21 20
+63 25 24
+25 63 26
+47 34 28
+34 47 35
+44 25 26
+46 11 51
+21 22 27
+63 24 23
+45 42 43
+42 45 46
+22 23 43
+23 22 63
+16 45 44
+16 26 63
+26 16 44
+16 63 62
+27 22 43
+27 28 38
+41 39 40
+39 41 36
+32 48 7
+32 6 31
+30 31 39
+31 30 32
+7 48 55
+54 49 50
+49 54 55
+35 51 50
+51 35 46
+43 44 45
+14 46 45
+34 35 50
+2 41 40
+28 39 38
+42 27 43
+42 46 47
+46 35 47
+9 55 54
+48 49 55
+54 50 53
+51 53 50
+53 51 52
+52 10 53
+53 9 54
